Make server object accessible in serverCallback
`userTextmessage()` and other functions in the `serverCallback` class need access to the current server object in order to be able to send messages, check permissions, etc. This was previously the case, but the server object was replaced with an `allianceauthenticatorApp` (app) object in a tidy of the codebase. This puts a server object back in scope while keeping the `app` object.

Signed-off-by: Eingang <eingang.vulpine@gmail.com>