Better formal arg names for userTextMessage()
`p`, which represents a `User` object isn't a great argument name. Changed to to `user` and "!kicktemps" command code adjusted so that the `users` loop variable  is `auser` instead so as not to conflict.

This argument naming change matches previous changes made to the corresponding formal argument for other user-related methods, like `userConnected()` and `userDisconnected`.

Some minor code white-spacing adjustments.

Signed-off-by: Eingang <eingang.vulpine@gmail.com>